♻️ extracted ActionTypes into proptypes
this avoids a double export in ActionButtons which messes
with React autoreload which requires single component exports
